feat: `mlg render` creates a dynamic webpage
Previously, `iframes` were used to hold subpages, and each
`.math` file had its own `.html` file.  Now, there is only a
single `index.html` that lists entities dynamically using
JavaScript.
