+++ src/main/kotlin/mathlingua/Main.kt
@@ -34,6 +34,7 @@ import java.nio.file.WatchService
 import kotlin.system.exitProcess
 import kotlinx.coroutines.runBlocking
 import mathlingua.backend.BackEnd
+import mathlingua.backend.SourceCollection
 import mathlingua.backend.SourceFile
 import mathlingua.backend.ValueSourceTracker
 import mathlingua.backend.isMathLinguaFile
@@ -336,41 +337,28 @@ private fun generateSignatureToPathJsCode(cwd: File, contentDir: File): String {
         signatureToPathBuilder.append("sigToPath.set('${entry.key}', '${entry.value}');")
     }
     signatureToPathBuilder.append("return sigToPath;")
-
-    return """
-        const SIGNATURE_TO_PATH = (function() {
-            $signatureToPathBuilder
-        })();
-
-        function mathlinguaToggleDropdown(id) {
-            const el = document.getElementById(id);
-            if (el) {
-                if (el.className === 'mathlingua-dropdown-menu-hidden') {
-                    el.className = 'mathlingua-dropdown-menu-shown';
-                } else {
-                    el.className = 'mathlingua-dropdown-menu-hidden';
-                }
-            }
-        }
-
-        function mathlinguaViewSignature(signature, id) {
-            mathlinguaToggleDropdown(id);
-            const path = SIGNATURE_TO_PATH.get(signature);
-            if (path) {
-                window.open(path, '_blank');
-            }
-        }
-    """.trimIndent()
+    return signatureToPathBuilder.toString()
 }
 
-private fun writeIndexFile(cwd: File, docsDir: File, contentDir: File): File {
+private fun sanitizeHtmlForJs(html: String) =
+    html.replace("\\", "\\\\").replace("\n", "\\n").replace("\"", "\\\"")
+
+private fun writeIndexFile(
+    files: List<String>,
+    cwd: File,
+    docsDir: File,
+    contentDir: File,
+    html: Boolean,
+    noexpand: Boolean,
+    errors: MutableList<ValueSourceTracker<ParseError>>
+): File {
     val allFileIds = mutableListOf<String>()
-    val builder = StringBuilder()
+    val fileListBuilder = StringBuilder()
     if (cwd.isDirectory) {
         val children = cwd.listFiles()
         if (children != null) {
             for (child in children) {
-                buildFileList(cwd, child, 0, builder, allFileIds)
+                buildFileList(cwd, child, 0, fileListBuilder, allFileIds)
             }
         }
     }
@@ -383,27 +371,54 @@ private fun writeIndexFile(cwd: File, docsDir: File, contentDir: File): File {
         val word = words[i]
         val pathToIndices = searchIndex[word]
         if (pathToIndices != null) {
-            searchIndexBuilder.append("const map$i = new Map();\n")
+            searchIndexBuilder.append("                const map$i = new Map();\n")
             val paths = pathToIndices.keys.toList()
             for (j in paths.indices) {
                 val path = paths[j]
                 val ids = pathToIndices[path]
                 if (ids != null) {
-                    searchIndexBuilder.append("const set${i}_$j = new Set();\n")
+                    searchIndexBuilder.append("                const set${i}_$j = new Set();\n")
                     for (id in ids) {
-                        searchIndexBuilder.append("set${i}_$j.add($id);\n")
+                        searchIndexBuilder.append("                set${i}_$j.add($id);\n")
                     }
-                    searchIndexBuilder.append("map$i.set('$path', set${i}_$j);\n")
+                    searchIndexBuilder.append("                map$i.set('$path', set${i}_$j);\n")
                 }
             }
-            searchIndexBuilder.append("index.set('$word', map$i);\n")
+            searchIndexBuilder.append("                index.set('$word', map$i);\n")
         }
     }
-    searchIndexBuilder.append("return index;")
+    searchIndexBuilder.append("                return index;")
 
-    val indexHtml = getIndexHtml(builder.toString(), searchIndexBuilder.toString(), allFileIds)
-    val indexFile = File(docsDir, "index.html")
-    indexFile.writeText(indexHtml)
+    val sigToPathCode = generateSignatureToPathJsCode(cwd, contentDir)
+
+    val sourceCollection = newSourceCollectionFromFiles(listOf(cwd))
+    val filesToProcess = mutableListOf<File>()
+    val inputFiles =
+        if (files.isEmpty()) {
+            listOf(cwd)
+        } else {
+            files.map { File(it) }
+        }
+    for (f in inputFiles) {
+        if (f.isDirectory) {
+            filesToProcess.addAll(f.walk().filter { isMathLinguaFile(it) })
+        } else if (isMathLinguaFile(f)) {
+            filesToProcess.add(f)
+        }
+    }
+
+    val pathToEntityMap =
+        generatePathToEntityList(cwd, filesToProcess, sourceCollection, html, noexpand, errors)
+    val pathToEntityBuilder = StringBuilder()
+    pathToEntityBuilder.append("                const map = new Map();\n")
+    val keys = pathToEntityMap.keys.toList()
+    for (path in keys) {
+        val key = path.replace("\"", "\\\"")
+        val entities = pathToEntityMap[key]!!.map { sanitizeHtmlForJs(it) }.map { "\"$it\"" }
+        pathToEntityBuilder.append("                map.set(\"$key\", $entities);\n")
+    }
+    pathToEntityBuilder.append("                return map;\n")
+
     val homeContentFile = File(cwd, "docs-home.html")
     val homeContent =
         if (homeContentFile.exists()) {
@@ -411,9 +426,46 @@ private fun writeIndexFile(cwd: File, docsDir: File, contentDir: File): File {
         } else {
             "<p>Create a file called <code>docs-home.html</code> to describe this repository.</p>"
         }
-    val homeFile = File(docsDir, "home.html")
-    homeFile.writeText(getHomeHtml(homeContent))
+    val homeHtml =
+        sanitizeHtmlForJs(
+            "<div style=\"font-size: 80%;font-family: Georgia, 'Times New Roman', Times, serif;\"><div class='mathlingua-top-level'>$homeContent</div></div>")
+    val indexHtml =
+        getIndexHtml(
+            fileListBuilder.toString(),
+            homeHtml,
+            searchIndexBuilder.toString(),
+            allFileIds,
+            sigToPathCode,
+            pathToEntityBuilder.toString())
+    val indexFile = File(docsDir, "index.html")
+    indexFile.writeText(indexHtml)
     return indexFile
+}
+
+private fun generatePathToEntityList(
+    cwd: File,
+    filesToProcess: List<File>,
+    sourceCollection: SourceCollection,
+    html: Boolean,
+    noexpand: Boolean,
+    errors: MutableList<ValueSourceTracker<ParseError>>
+): Map<String, List<String>> {
+    val result = mutableMapOf<String, List<String>>()
+    for (f in filesToProcess) {
+        val path = f.relativePath(cwd)
+        val pair = sourceCollection.prettyPrint(file = f, html = html, doExpand = !noexpand)
+        errors.addAll(
+            pair.second.map {
+                ValueSourceTracker(
+                    value = it,
+                    source =
+                        SourceFile(
+                            file = f, content = "", validation = validationFailure(emptyList())),
+                    tracker = null)
+            })
+        result[path] = pair.first
+    }
+    return result
 }
 
 private fun render(
@@ -443,38 +495,11 @@ private fun render(
     contentDir.mkdirs()
 
     val html = format == "html"
-    if (html && !stdout) {
-        val indexFile = writeIndexFile(cwd, docsDir, contentDir)
-        log("Wrote ${indexFile.relativeTo(cwd)}")
-    }
-
-    val sigToPathCode = generateSignatureToPathJsCode(cwd, contentDir)
-
     val errors = mutableListOf<ValueSourceTracker<ParseError>>()
-    for (f in filesToProcess) {
-        val pair =
-            sourceCollection.prettyPrint(
-                file = f, html = html, js = sigToPathCode, doExpand = !noexpand)
-        errors.addAll(
-            pair.second.map {
-                ValueSourceTracker(
-                    value = it,
-                    source =
-                        SourceFile(
-                            file = f, content = "", validation = validationFailure(emptyList())),
-                    tracker = null)
-            })
-        val ext =
-            if (html) {
-                ".html"
-            } else {
-                ".out.math"
-            }
-        val docRelFile = File(docsDir, f.relativePath(cwd))
-        val docRelParent = docRelFile.parentFile
-        val docRelName = docRelFile.nameWithoutExtension + ext
-        val outFile = File(docRelParent, docRelName)
-        write(cwd = cwd, content = pair.first, outFile = outFile, stdout = stdout)
+
+    if (html && !stdout) {
+        val indexFile = writeIndexFile(files, cwd, docsDir, contentDir, html, noexpand, errors)
+        log("Wrote ${indexFile.relativeTo(cwd)}")
     }
 
     if (!stdout) {
@@ -672,17 +697,617 @@ fun main(args: Array<String>) {
     mlg.main(args)
 }
 
-fun getIndexHtml(fileListHtml: String, searchIndexInitCode: String, allFileIds: List<String>) =
+fun getIndexHtml(
+    fileListHtml: String,
+    homeHtml: String,
+    searchIndexInitCode: String,
+    allFileIds: List<String>,
+    signatureToPathCode: String,
+    pathToEntityList: String
+) =
     """
+<!DOCTYPE html>
 <html>
     <head>
-        <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0">
+        <meta name="viewport" content="width=100%, initial-scale=1.0">
         <meta content="text/html;charset=utf-8" http-equiv="Content-Type">
         <meta content="utf-8" http-equiv="encoding">
+        <link rel="stylesheet"
+              href="https://cdn.jsdelivr.net/npm/katex@0.11.1/dist/katex.min.css"
+              integrity="sha384-zB1R0rpPzHqg7Kpt0Aljp8JPLqbXI3bhnPWROx27a9N0Ll6ZP/+DiW/UqRcLbRjq"
+              crossorigin="anonymous">
+        <script defer
+                src="https://cdn.jsdelivr.net/npm/katex@0.11.1/dist/katex.min.js"
+                integrity="sha384-y23I5Q6l+B6vatafAwxRu/0oK/79VlbSz7Q9aiSZUvyWYIYsd+qj+o24G5ZU2zJz"
+                crossorigin="anonymous">
+        </script>
+        <script>
+            const ALL_FILE_IDS = [${allFileIds.joinToString(",") { "'$it'" }}];
+            ALL_FILE_IDS.push('home');
+            const INITIAL_SRC = "home.html";
+            const SEARCH_INDEX = (function() {
+                $searchIndexInitCode
+            })();
+
+            const SIGNATURE_TO_PATH = (function() {
+                $signatureToPathCode
+            })();
+
+            const PATH_TO_ENTITY_LIST = (function() {
+                $pathToEntityList
+            })();
+
+            const HOME_SRC = "$homeHtml";
+
+            function mathlinguaToggleDropdown(id) {
+                const el = document.getElementById(id);
+                if (el) {
+                    if (el.className === 'mathlingua-dropdown-menu-hidden') {
+                        el.className = 'mathlingua-dropdown-menu-shown';
+                    } else {
+                        el.className = 'mathlingua-dropdown-menu-hidden';
+                    }
+                }
+            }
+
+            function mathlinguaViewSignature(signature, id) {
+                mathlinguaToggleDropdown(id);
+                const path = SIGNATURE_TO_PATH.get(signature);
+                if (path) {
+                    window.open(path, '_blank');
+                }
+            }
+
+            function buildMathFragment(rawText) {
+                var text = rawText;
+                if (text[0] === '"') {
+                    text = text.substring(1);
+                }
+                if (text[text.length - 1] === '"') {
+                    text = text.substring(0, text.length - 1);
+                }
+                text = text.replace(/([a-zA-Z0-9])\?\??/g, '${'$'}1');
+                const fragment = document.createDocumentFragment();
+                var buffer = '';
+                var i = 0;
+                while (i < text.length) {
+                    if (text[i] === '\\' && text[i+1] === '[') {
+                        i += 2; // skip over \ and [
+                        fragment.appendChild(document.createTextNode(buffer));
+                        buffer = '';
+
+                        const span = document.createElement('span');
+                        var math = '';
+                        while (i < text.length &&
+                            !(text[i] === '\\' && text[i+1] === ']')) {
+                            math += text[i++];
+                        }
+                        if (text[i] === '\\') {
+                            i++; // move past the \
+                        }
+                        if (text[i] === ']') {
+                            i++; // move past the ]
+                        }
+                        try {
+                            katex.render(math, span, {
+                                throwOnError: true,
+                                displayMode: true
+                            });
+                        } catch {
+                            span.appendChild(document.createTextNode(math));
+                        }
+                        fragment.appendChild(span);
+                    } else if (text[i] === '\\' && text[i+1] === '(') {
+                        i += 2; // skip over \ and ()
+                        fragment.appendChild(document.createTextNode(buffer));
+                        buffer = '';
+
+                        const span = document.createElement('span');
+                        var math = '';
+                        while (i < text.length &&
+                            !(text[i] === '\\' && text[i+1] === ')')) {
+                            math += text[i++];
+                        }
+                        if (text[i] === '\\') {
+                            i++; // move past the \
+                        }
+                        if (text[i] === ')') {
+                            i++; // move past the )
+                        }
+                        try {
+                            katex.render(math, span, {
+                                throwOnError: true,
+                                displayMode: true
+                            });
+                        } catch {
+                            span.appendChild(document.createTextNode(math));
+                        }
+                        fragment.appendChild(span);
+                    } else if (text[i] === '${'$'}' && text[i+1] === '${'$'}') {
+                        i += 2; // skip over ${'$'} and ${'$'}
+                        fragment.appendChild(document.createTextNode(buffer));
+                        buffer = '';
+
+                        const span = document.createElement('span');
+                        var math = '';
+                        while (i < text.length &&
+                            !(text[i] === '${'$'}' && text[i+1] === '${'$'}')) {
+                            math += text[i++];
+                        }
+                        if (text[i] === '${'$'}') {
+                            i++; // move past the ${'$'}
+                        }
+                        if (text[i] === '${'$'}') {
+                            i++; // move past the ${'$'}
+                        }
+                        try {
+                            katex.render(math, span, {
+                                throwOnError: true,
+                                displayMode: true
+                            });
+                        } catch {
+                            span.appendChild(document.createTextNode(math));
+                        }
+                        fragment.appendChild(span);
+                    } else if (text[i] === '${'$'}') {
+                        i++; // skip over the ${'$'}
+                        fragment.appendChild(document.createTextNode(buffer));
+                        buffer = '';
+
+                        const span = document.createElement('span');
+                        var math = '';
+                        while (i < text.length &&
+                             text[i] !== '${'$'}') {
+                            math += text[i++];
+                        }
+                        if (text[i] === '${'$'}') {
+                            i++; // move past the ${'$'}
+                        }
+                        try {
+                            katex.render(math, span, {
+                                throwOnError: true,
+                                displayMode: true
+                            });
+                        } catch {
+                            span.appendChild(document.createTextNode(math));
+                        }
+                        fragment.appendChild(span);
+                    } else {
+                        buffer += text[i++];
+                    }
+                }
+
+                if (buffer.length > 0) {
+                    fragment.appendChild(document.createTextNode(buffer));
+                }
+
+                return fragment;
+            }
+
+            function render(node) {
+                if (node.className && node.className.indexOf('no-render') >= 0) {
+                    return;
+                }
+
+                let isInWritten = false;
+                const parent = node.parentNode;
+                if (node.className === 'mathlingua') {
+                    for (let i=0; i<node.childNodes.length; i++) {
+                        const n = node.childNodes[i];
+                        if (n && n.className === 'mathlingua-header' &&
+                            n.textContent === 'written:') {
+                            isInWritten = true;
+                            break;
+                        }
+                    }
+                }
+
+                for (let i = 0; i < node.childNodes.length; i++) {
+                    const child = node.childNodes[i];
+
+                    // node is an element node => nodeType === 1
+                    // node is an attribute node => nodeType === 2
+                    // node is a text node => nodeType === 3
+                    // node is a comment node => nodeType === 8
+                    if (child.nodeType === 3) {
+                        let text = child.textContent;
+                        if (text.trim()) {
+                            if (isInWritten) {
+                                // if the text is in a written: section
+                                // turn "some text" to \[some text\]
+                                // so the text is in math mode
+                                if (text[0] === '"') {
+                                    text = text.substring(1);
+                                }
+                                if (text[text.length - 1] === '"') {
+                                    text = text.substring(0, text.length - 1);
+                                }
+                                text = '\\[' + text + '\\]';
+                            }
+                            const fragment = buildMathFragment(text);
+                            i += fragment.childNodes.length - 1;
+                            node.replaceChild(fragment, child);
+                        }
+                    } else if (child.nodeType === 1) {
+                        render(child);
+                    }
+                }
+            }
+
+            function setup() {
+                render(document.body);
+                const params = new URLSearchParams(window.location.search);
+                const showIds = new Set(params.getAll('show'));
+                if (showIds.size > 0) {
+                    let i = 0;
+                    while (true) {
+                        const id = '' + (i++);
+                        const el = document.getElementById(id);
+                        if (!el) {
+                            break;
+                        }
+                        if (showIds.has(id)) {
+                            el.style.display = 'block';
+                        } else {
+                            el.style.display = 'none';
+                        }
+                    }
+                }
+            }
+
+            function forMobile() {
+                return window?.screen?.width <= 500;
+            }
+
+            let open = !forMobile();
+
+            function toggleSidePanel() {
+                if (open) {
+                    document.getElementById('sidebar').style.width = '0';
+                    document.getElementById('main').style.marginLeft = '0';
+                } else {
+                    let margin = forMobile() ? '50%' : '20%';
+                    document.getElementById('sidebar').style.width = margin;
+                    document.getElementById('main').style.marginLeft = margin;
+                }
+                open = !open;
+            }
+
+            function view(path) {
+                const content = document.getElementById('__main_content__');
+                if (content) {
+                    for (const path of ALL_FILE_IDS) {
+                        if (!path) {
+                            continue;
+                        }
+                        const el = document.getElementById(path);
+                        if (el) {
+                            el.style.fontStyle = 'normal';
+                        }
+                    }
+                    const id = path.replace(/\.html.*/, '');
+                    if (id) {
+                        const selectedEntry = document.getElementById(id);
+                        if (selectedEntry) {
+                            selectedEntry.style.fontStyle = 'italic';
+                        }
+                    }
+
+                    if (!path || path === 'home.html') {
+                        content.innerHTML = HOME_SRC;
+                        return;
+                    }
+
+                    let idSet = null;
+                    const parts = path.split(';');
+                    if (parts.length === 2) {
+                        idSet = new Set(parts[1].split(','));
+                    }
+
+                    const newPath = path.replace(/;.*/g, '').replace(/\.html${'$'}/, ".math");
+                    const entityList = PATH_TO_ENTITY_LIST.get(newPath);
+                    if (entityList) {
+                        while (content.firstChild) {
+                            content.removeChild(content.firstChild);
+                        }
+                        for (let i=0; i<entityList.length; i++) {
+                            if (!!idSet && idSet.has(id)) {
+                                continue;
+                            }
+
+                            const el = document.createElement('div');
+                            el.className = 'mathlingua-top-level';
+                            el.innerHTML = entityList[i];
+                            content.appendChild(el);
+                            content.appendChild(document.createElement('br'));
+                            content.appendChild(document.createElement('br'));
+                        }
+                        render(content);
+                    }
+                }
+            }
+
+            function clearSearch() {
+                for (const id of ALL_FILE_IDS) {
+                    if (!id) {
+                        continue;
+                    }
+                    const el = document.getElementById(id);
+                    if (el) {
+                        el.style.display = 'block';
+                        el.setAttribute('onclick', "view('" + id + ".html" + "')");
+                    }
+                }
+                view(INITIAL_SRC);
+                const el = document.getElementById('search-input');
+                if (el) {
+                    el.value = '';
+                }
+            }
+
+            function search() {
+                const el = document.getElementById('search-input');
+                if (el) {
+                    if (!el.value || !el.value.trim()) {
+                        clearSearch();
+                        return;
+                    }
+
+                    const terms = el.value.split(' ')
+                                    .map(it => it.trim().toLowerCase())
+                                    .filter(it => it.length > 0);
+                    if (terms.length == 0) {
+                        clearSearch();
+                        return;
+                    }
+
+                    view('');
+                    const pathsToIndices = new Map();
+                    if (terms.length > 0) {
+                        const temp = SEARCH_INDEX.get(terms[0]) || new Map();
+                        for (const [path, indices] of temp) {
+                            pathsToIndices.set(path, new Set(indices));
+                        }
+                    }
+
+                    for (let i=1; i<terms.length; i++) {
+                        const term = terms[i];
+                        const temp = SEARCH_INDEX.get(term) || new Map();
+                        for (const path of pathsToIndices.keys()) {
+                            if (!temp.has(path)) {
+                                pathsToIndices.delete(path);
+                            } else {
+                                const intersection = new Set();
+                                const tempIndices = temp.get(path) || new Set();
+                                for (const id of (pathsToIndices.get(path) || new Set())) {
+                                    if (tempIndices.has(id)) {
+                                        intersection.add(id);
+                                    }
+                                }
+                                if (intersection.size > 0) {
+                                    pathsToIndices.set(path, intersection);
+                                } else {
+                                    pathsToIndices.delete(path);
+                                }
+                            }
+                        }
+                    }
+
+                    if (pathsToIndices.size === 0) {
+                        alert('No results found');
+                        return;
+                    }
+
+                    const pathToNewPath = new Map();
+                    let firstPath = null;
+                    for (const [path, ids] of pathsToIndices) {
+                        let newPath = path + '.html;' + Array.from(ids).toString();
+                        if (firstPath === null) {
+                            firstPath = newPath;
+                        }
+                        pathToNewPath.set(path, newPath);
+                    }
+
+                    for (const id of ALL_FILE_IDS) {
+                        if (!id) {
+                            continue;
+                        }
+                        const el = document.getElementById(id);
+                        if (el) {
+                            if (pathToNewPath.has(id)) {
+                                el.style.display = 'block';
+                                el.setAttribute('onclick', "view('" + pathToNewPath.get(id) + "')");
+                            } else if (id !== 'home') {
+                                el.style.display = 'none';
+                            }
+                        }
+                    }
+
+                    for (const id of ALL_FILE_IDS) {
+                        if (pathToNewPath.has(id)) {
+                            const parts = id.split('/');
+                            while (parts.length > 0) {
+                                parts.pop();
+                                let parent = parts.join('/');
+                                if (parent) {
+                                    const parentEl = document.getElementById(parent);
+                                    if (parentEl) {
+                                        parentEl.style.display = 'block';
+                                    }
+                                }
+                            }
+                        }
+                    }
+
+                    if (firstPath) {
+                        view(firstPath);
+                    }
+                }
+            }
+
+            function initPage() {
+                const el = document.getElementById('search-input');
+                if (el) {
+                    el.addEventListener("keyup", function(event) {
+                        if (event.keyCode === 13) {
+                            event.preventDefault();
+                            search();
+                        }
+                    });
+                }
+                setup();
+                view('');
+            }
+        </script>
         <style>
+            .content {
+                margin-top: 1.5em;
+                margin-bottom: 1em;
+                font-size: 1em;
+                width: 70%;
+                margin-left: auto; /* for centering content */
+                margin-right: auto; /* for centering content */
+            }
+
+            @media screen and (max-width: 400px) {
+                .content {
+                    width: 90%;
+                }
+            }
+
             body {
                 background-color: #fafafa;
-                overflow: hidden;
+                padding-bottom: 1em;
+            }
+
+            hr {
+                border: 0.5px solid #efefef;
+            }
+
+            h1, h2, h3, h4 {
+                color: #0055bb;
+            }
+
+            .mathlingua-top-level {
+                background-color: white;
+                border: solid;
+                border-width: 1px;
+                border-color: rgba(200, 200, 200);
+                border-radius: 2px;
+                box-shadow: rgba(0, 0, 0, 0.5) 0px 3px 10px,
+                            inset 0  0 10px 0 rgba(200, 200, 200, 0.25);
+                padding-top: 1.25em;
+                padding-bottom: 1em;
+                padding-left: 1.1em;
+                padding-right: 1.1em;
+                max-width: 90%;
+                width: max-content;
+                margin-left: auto; /* for centering content */
+                margin-right: auto; /* for centering content */
+            }
+
+            .end-mathlingua-top-level {
+                padding-top: 0.5em;
+                margin: 0;
+            }
+
+            .mathlingua {
+                font-family: monospace;
+            }
+
+            .mathlingua-header {
+                color: #0055bb;
+                text-shadow: 0px 1px 0px rgba(255,255,255,0), 0px 0.4px 0px rgba(0,0,0,0.2);
+                font-size: 80%;
+                font-family: Georgia, 'Times New Roman', Times, serif;
+            }
+
+            .mathlingua-whitespace {
+                padding: 0;
+                margin: 0;
+                margin-left: 1ex;
+            }
+
+            .mathlingua-id {
+                color: #5500aa;
+                overflow-x: scroll;
+            }
+
+            .mathlingua-text {
+                color: #000000;
+                display: block;
+                margin: 0 0 -1em 0;
+                padding: 0 0 0 2.5em;
+                font-size: 80%;
+                font-family: Georgia, 'Times New Roman', Times, serif;
+            }
+
+            .mathlingua-text-no-render {
+                color: #000000;
+                display: block;
+                margin: 0 0 -1em 0;
+                padding: 0 0 0 2.5em;
+                font-size: 80%;
+                font-family: Georgia, 'Times New Roman', Times, serif;
+            }
+
+            .mathlingua-url {
+                color: #000000;
+                display: block;
+                margin: 0 0 -1em 0;
+                padding: 0 0 0 2.5em;
+                font-size: 80%;
+                font-family: Georgia, 'Times New Roman', Times, serif;
+            }
+
+            .mathlingua-statement-no-render {
+                color: #007377;
+            }
+
+            .mathlingua-statement-container {
+                display: inline;
+            }
+
+            .mathlingua-dropdown-menu-shown {
+                position: absolute;
+                display: block;
+                z-index: 1;
+                background-color: #ffffff;
+                box-shadow: rgba(0, 0, 0, 0.5) 0px 3px 10px,
+                            inset 0  0 10px 0 rgba(200, 200, 200, 0.25);
+                border: solid;
+                border-width: 1px;
+                border-radius: 0px;
+                border-color: rgba(200, 200, 200);
+            }
+
+            .mathlingua-dropdown-menu-hidden {
+                display: none;
+            }
+
+            .mathlingua-dropdown-menu-item {
+                display: block;
+                margin: 0.75ex;
+            }
+
+            .mathlingua-dropdown-menu-item:hover {
+                font-style: italic;
+            }
+
+            .katex {
+                font-size: 0.9em;
+            }
+
+            .katex-display {
+                display: contents;
+            }
+
+            .katex-display > .katex {
+                display: contents;
+            }
+
+            .katex-display > .katex > .katex-html {
+                display: contents;
             }
 
             .topbar {
@@ -795,209 +1420,7 @@ fun getIndexHtml(fileListHtml: String, searchIndexInitCode: String, allFileIds: 
                     margin-left: 0;
                 }
             }
-
-            #__content__frame__ {
-                border: none;
-                width: 100%;
-                height: 100%;
-                margin: 0;
-                padding: 0;
-            }
-
-            hr {
-                border: 0.5px solid #efefef;
-            }
         </style>
-        <script>
-            const ALL_FILE_IDS = [${allFileIds.joinToString(",") { "'$it'" }}];
-            ALL_FILE_IDS.push('home');
-            const INITIAL_SRC = "home.html";
-            const SEARCH_INDEX = (function() {
-                $searchIndexInitCode
-            })();
-
-            function forMobile() {
-                return window?.screen?.width <= 500;
-            }
-
-            let open = !forMobile();
-
-            function toggleSidePanel() {
-                if (open) {
-                    document.getElementById('sidebar').style.width = '0';
-                    document.getElementById('main').style.marginLeft = '0';
-                } else {
-                    let margin = forMobile() ? '50%' : '20%';
-                    document.getElementById('sidebar').style.width = margin;
-                    document.getElementById('main').style.marginLeft = margin;
-                }
-                open = !open;
-            }
-
-            function view(path) {
-                const content = document.getElementById('__content__frame__');
-                if (content) {
-                    content.src = path;
-                    for (const path of ALL_FILE_IDS) {
-                        if (!path) {
-                            continue;
-                        }
-                        const el = document.getElementById(path);
-                        if (el) {
-                            el.style.fontStyle = 'normal';
-                        }
-                    }
-                    const id = path.replace(/\.html.*/, '');
-                    if (id) {
-                        const selectedEntry = document.getElementById(id);
-                        if (selectedEntry) {
-                            selectedEntry.style.fontStyle = 'italic';
-                        }
-                    }
-                }
-            }
-
-            function clearSearch() {
-                for (const id of ALL_FILE_IDS) {
-                    if (!id) {
-                        continue;
-                    }
-                    const el = document.getElementById(id);
-                    if (el) {
-                        el.style.display = 'block';
-                        el.setAttribute('onclick', "view('" + id + ".html" + "')");
-                    }
-                }
-                view(INITIAL_SRC);
-                const el = document.getElementById('search-input');
-                if (el) {
-                    el.value = '';
-                }
-            }
-
-            function search(terms) {
-                const el = document.getElementById('search-input');
-                if (el) {
-                    if (!el.value || !el.value.trim()) {
-                        clearSearch();
-                        return;
-                    }
-
-                    const terms = el.value.split(' ')
-                                    .map(it => it.trim().toLowerCase())
-                                    .filter(it => it.length > 0);
-                    if (terms.length == 0) {
-                        clearSearch();
-                        return;
-                    }
-
-                    view('');
-                    const pathsToIndices = new Map();
-                    if (terms.length > 0) {
-                        const temp = SEARCH_INDEX.get(terms[0]) || new Map();
-                        for (const [path, indices] of temp) {
-                            pathsToIndices.set(path, new Set(indices));
-                        }
-                    }
-
-                    for (let i=1; i<terms.length; i++) {
-                        const term = terms[i];
-                        const temp = SEARCH_INDEX.get(term) || new Map();
-                        for (const path of pathsToIndices.keys()) {
-                            if (!temp.has(path)) {
-                                pathsToIndices.delete(path);
-                            } else {
-                                const intersection = new Set();
-                                const tempIndices = temp.get(path) || new Set();
-                                for (const id of (pathsToIndices.get(path) || new Set())) {
-                                    if (tempIndices.has(id)) {
-                                        intersection.add(id);
-                                    }
-                                }
-                                if (intersection.size > 0) {
-                                    pathsToIndices.set(path, intersection);
-                                } else {
-                                    pathsToIndices.delete(path);
-                                }
-                            }
-                        }
-                    }
-
-                    if (pathsToIndices.size === 0) {
-                        alert('No results found');
-                        return;
-                    }
-
-                    const pathToNewPath = new Map();
-                    let firstPath = null;
-                    for (const [path, ids] of pathsToIndices) {
-                        let newPath = path + '.html';
-                        if (ids.size > 0) {
-                            newPath += '?';
-                            let isFirst = true;
-                            for (const id of ids) {
-                                if (!isFirst) {
-                                    newPath += '&'
-                                }
-                                isFirst = false;
-                                newPath += "show=" + id;
-                            }
-                            if (!firstPath) {
-                                firstPath = newPath;
-                            }
-                            pathToNewPath.set(path, newPath);
-                        }
-                    }
-
-                    for (const id of ALL_FILE_IDS) {
-                        if (!id) {
-                            continue;
-                        }
-                        const el = document.getElementById(id);
-                        if (el) {
-                            if (pathToNewPath.has(id)) {
-                                el.style.display = 'block';
-                                el.setAttribute('onclick', "view('" + pathToNewPath.get(id) + "')");
-                            } else {
-                                el.style.display = 'none';
-                            }
-                        }
-                    }
-
-                    for (const id of ALL_FILE_IDS) {
-                        if (pathToNewPath.has(id)) {
-                            const parts = id.split('/');
-                            while (parts.length > 0) {
-                                parts.pop();
-                                let parent = parts.join('/');
-                                if (parent) {
-                                    const parentEl = document.getElementById(parent);
-                                    if (parentEl) {
-                                        parentEl.style.display = 'block';
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    if (firstPath) {
-                        view(firstPath);
-                    }
-                }
-            }
-
-            function initPage() {
-                const el = document.getElementById('search-input');
-                if (el) {
-                    el.addEventListener("keyup", function(event) {
-                        if (event.keyCode === 13) {
-                            event.preventDefault();
-                            search();
-                        }
-                    });
-                }
-            }
-        </script>
     </head>
     <body id="main" onload="initPage()">
         <div id="top-bar" class="topbar">
@@ -1015,32 +1438,8 @@ fun getIndexHtml(fileListHtml: String, searchIndexInitCode: String, allFileIds: 
             $fileListHtml
         </div>
 
-        <iframe id="__content__frame__" src="home.html"></iframe>
-    </body>
-</html>
-"""
-
-fun getHomeHtml(content: String) =
-    """
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta name="viewport" content="width=100%, initial-scale=1.0">
-        <meta content="text/html;charset=utf-8" http-equiv="Content-Type">
-        <meta content="utf-8" http-equiv="encoding">
-        <style>
-            body {
-                font-size: 80%;
-                font-family: Georgia, 'Times New Roman', Times, serif;
-            }
-
-            h1, h2, h3, h4 {
-                color: #0055bb;
-            }
-        </style>
-    </head>
-    <body>
-        $content
+        <div class="content" id="__main_content__">
+        </div>
     </body>
 </html>
 """
